NSUndoManagerWillCloseUndoGroupNotification no post if no changes; -hasChanges; fix availability macro.
TOPIC 1  I think we did not get
NSUndoManagerWillCloseUndoGroupNotification quite correct.  Graham's
comment in -endUndoGrouping says:  "closing outer group. If it's empty,
remove it. This is what NSUndoManager should do, but doesn't. That means
that this um is easier to use in some situations such as grouping across
a series of events that may or may not submit undo tasks.  If no tasks
were submitted, no bogus empty undo task remains on the stack. In
addition no closure notification is sent so as far as the client is
concerned, it just never happened."  Very good.  But we did not follow
that pattern when we added posting of the new
NSUndoManagerDidCloseUndoGroupNotification.  As a matter of fact, if I
open an undo group, then close it without adding any tasks, the
NSUndoManagerWillCloseUndoGroupNotification does *not* get posted, but
the NSUndoManagerDidCloseUndoGroupNotification *does* get posted.
Clearly that does not make sense.  If I remember correctly, NSDocument
in Mac OS 10.7 listens for the DidClose notification while in previous
OS X versions listened for the WillClose notification.  So we've
effectively reverted to the old NSUndoManager behavior in 10.7.  Indeed,
in my Core Data : NSPersistenDocument app, when I open+close an empty
undo group, …  • if I've *not* enabled autosaves-in-place, the
notification posting causes -updateChangeCount: to be sent to my
document, which dirties the red 'close' button, even though the document
is not really dirty.   • if I *have* enabled autosaves-in-place, then
the next time that autosave is triggered (the easiest way to make this
happen is to activate a different app), -[NSDocument
writeSafelyToURL:ofType:forSaveOperation:error:] is invoked.  (Last
week, I was seeing the file's modification date get touched, which was
really bad, but today, for some reason, I don't.)  However, I've come to
feel that the fewer machinations get triggered by autosave, not to
mention performance, the less trouble there's likely to be in the world
;)  I've confirmed that the above behaviors are the same in Apple's
ancient (2007) "Department and Employees" Sample Project, which still
compiles in 10.7 SDK albeit with a lot of warnings.  (After fixing, with
autosaves-in-place, then -[NSDocument
autosaveWithImplicitCancellability:completionHandler:] still gets
invoked when activating a different app, but the call chain is stopped
before it invokes -writeSafelyToURL::::.  Probably the clean/dirty
state, which still exists but is hidden in Lion, is checked in
-saveToURL:ofType:forSaveOperation:completionHandler:.)   *  *  *  TOPIC
2  In fixing that problem, I ran into another little issue.  If the
removed empty undo group was the only undo group, then [self
currentGroup] is nil, and the BOOL value of    [[self currentGroup]
isEmpty]  is NO, which is not what you'd expect.  When you ask a nil
group if it is empty, it should answer YES.  I don't know if this
affects the logic driving the *Will*Close notification, or other uses of
this method, but I don't like it.  So I negated everything and changed
-isEmpty to -hasTask.  This method was exposed in the interface, so it's
an API change, but only for GCUndoGroup.  NSUndoGroup, if there is such
a class, is Apple-private.  But since forgetting to check for nil is a
common programming error, I think -hasTask is better.  *  *  *  TOPIC 3
The availability macro we used is incorrect.  In this line of code, #if
MAC_OS_X_VERSION_MIN_REQUIRED >= 1070 which we use in two places,
MAC_OS_X_VERSION_MIN_REQUIRED is the minimum system version required to
*run* the application.  Its value is equal to the resolved value of the
"Mac OS X Deployment Target" (MACOSX_DEPLOYMENT_TARGET) build setting.
What we want is MAC_OS_X_VERSION_MAX_ALLOWED, which is the maximum
system version in which any *API that you use* must be available.  Its
value is equal to the resolved value of the "Base SDK" (SDKROOT) build
setting.